lastupdate change

+++ entity/lastUpdate.go
@@ -3,14 +3,14 @@ package entity
 import "time"
 
 type LastUpdateOnlineDb struct {
-	ID         uint `gorm:"primaryKey"`
-	UserID     uint `gorm:"unique"`
+	ID         uint   `gorm:"primaryKey"`
+	UserID     uint64 `gorm:"unique"`
 	TimeOnline time.Time
 }
 
 type LastTapUserDb struct {
-	ID       uint `gorm:"primaryKey"`
-	UserID   uint `gorm:"unique"`
+	ID       uint   `gorm:"primaryKey"`
+	UserID   uint64 `gorm:"unique"`
 	CountTap uint64
 	TimeTap  time.Time
 }
